First commit.
Implemented the Shunting Yard Algorithm (SYA) to parse the circuit expressions.
Implemented the Circuit class to store the values of the components and
to evaluate the result of the SYA. Need to add tests and fix some of the
design choices I have made. Still need to implement the module for the
best fit.